ReadMe-file for the project BankSimulator.
First of all we created an instance for the class LogIn. In the LogIn class you have to declare the Admin of the program. 
And when the Admin is created the instance of the Admin is automatically created.
When the Admin instance is done we will return to the LogIn class again.
In the LogIn class the user have 3 tries to login, otherwise the user is getting locked and need to be unlocked by the admin.
When we have logged in on the admin, we are able to unlock locked customers, create new customers or get customers info in the bank.
In the Admin class we handle the verification of the customers and also handle all the log in tries.
The Admin and the Customer class derives from the User class.

In the Customer Class you are able to open bank and saving accounts, apply for loans and transfer the money between their accounts.
They are also able to display all the accounts they have on the bank and the transfers between the accounts.
When the customer is trying to log in and fails 3 times with the password, the customer is getting locked out from the bank. To be unlocked you have to contact the admin to be unlocked. 
We had set the mortgage limit to 5 times the balance you have on the bank currently, and the interest rate will always be 30% no matter what.
When we are creating new bank and saving accounts we creates new objects. Each object is their own account, and these objects are connected to the Customer class.

You are only able to create 1 admin and able to create multiplies of customers. The customers are then able to create UNLIMITED bank and saving accounts, but customers are only able to apply to 1 loan MAXIMUM.

The Bank class is the class you use when you are logging in as a customer. It handles all your menu options that the customers has. The Bank class is the Main Menu for the Customers. The Bank class handles all the connections/communications for the Customer class.

- by Mohammed, Kristina and Erik - Team Citron.